backend/requirements.txt
# FASTAPI + SERVER
fastapi==0.115.5
uvicorn[standard]==0.32.0

# CONFIG / SETTINGS
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1

# DATABASE
sqlalchemy==2.0.36

# AUTH
passlib==1.7.4
bcrypt==3.2.2
python-jose==3.3.0
email-validator==2.2.0

# FILE UPLOAD / MISC
python-multipart==0.0.9

# IMAGE PROCESSING
Pillow==10.4.0
numpy==2.1.2
opencv-python==4.10.0.84

# MACHINE LEARNING
torch
torchvision

dominate
visdom
wandb

# CLOUDINARY
cloudinary